Update: logic 걸려있을때 dnd기능 분기처리
